+++ src/models/get_transaction_details_by_transaction_id_from_callback_ribsd.rs
@@ -16,7 +16,7 @@
 pub struct GetTransactionDetailsByTransactionIdFromCallbackRibsd {
     /// Represents the time at which a particular transaction can be added to the blockchain.
     #[serde(rename = "locktime")]
-    pub locktime: i32,
+    pub locktime: i64,
     /// Represents the total size of this transaction.
     #[serde(rename = "size")]
     pub size: i32,
@@ -33,7 +33,7 @@ pub struct GetTransactionDetailsByTransactionIdFromCallbackRibsd {
 
 impl GetTransactionDetailsByTransactionIdFromCallbackRibsd {
     /// Dash
-    pub fn new(locktime: i32, size: i32, version: i32, vin: Vec<crate::models::GetTransactionDetailsByTransactionIdFromCallbackRibsdVin>, vout: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vout>) -> GetTransactionDetailsByTransactionIdFromCallbackRibsd {
+    pub fn new(locktime: i64, size: i32, version: i32, vin: Vec<crate::models::GetTransactionDetailsByTransactionIdFromCallbackRibsdVin>, vout: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vout>) -> GetTransactionDetailsByTransactionIdFromCallbackRibsd {
         GetTransactionDetailsByTransactionIdFromCallbackRibsd {
             locktime,
             size,
