+++ src/models/get_transaction_details_by_transaction_id_from_callback_ribsd_vin.rs
@@ -22,7 +22,7 @@ pub struct GetTransactionDetailsByTransactionIdFromCallbackRibsdVin {
     pub script_sig: Box<crate::models::GetTransactionDetailsByTransactionIdribsd2ScriptSig>,
     /// Represents the script sequence number.
     #[serde(rename = "sequence")]
-    pub sequence: i32,
+    pub sequence: i64,
     /// String representation of the txid
     #[serde(rename = "txid", skip_serializing_if = "Option::is_none")]
     pub txid: Option<String>,
@@ -37,7 +37,7 @@ pub struct GetTransactionDetailsByTransactionIdFromCallbackRibsdVin {
 }
 
 impl GetTransactionDetailsByTransactionIdFromCallbackRibsdVin {
-    pub fn new(addresses: Vec<String>, script_sig: crate::models::GetTransactionDetailsByTransactionIdribsd2ScriptSig, sequence: i32, txinwitness: Vec<String>) -> GetTransactionDetailsByTransactionIdFromCallbackRibsdVin {
+    pub fn new(addresses: Vec<String>, script_sig: crate::models::GetTransactionDetailsByTransactionIdribsd2ScriptSig, sequence: i64, txinwitness: Vec<String>) -> GetTransactionDetailsByTransactionIdFromCallbackRibsdVin {
         GetTransactionDetailsByTransactionIdFromCallbackRibsdVin {
             addresses,
             coinbase: None,
