+++ src/models/get_transaction_details_by_transaction_idribsd2.rs
@@ -16,7 +16,7 @@
 pub struct GetTransactionDetailsByTransactionIdribsd2 {
     /// Represents the time at which a particular transaction can be added to the blockchain.
     #[serde(rename = "locktime")]
-    pub locktime: i32,
+    pub locktime: i64,
     /// Represents the total size of this transaction.
     #[serde(rename = "size")]
     pub size: i32,
@@ -33,7 +33,7 @@ pub struct GetTransactionDetailsByTransactionIdribsd2 {
 
 impl GetTransactionDetailsByTransactionIdribsd2 {
     /// Dash
-    pub fn new(locktime: i32, size: i32, version: i32, vin: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vin>, vout: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vout>) -> GetTransactionDetailsByTransactionIdribsd2 {
+    pub fn new(locktime: i64, size: i32, version: i32, vin: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vin>, vout: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vout>) -> GetTransactionDetailsByTransactionIdribsd2 {
         GetTransactionDetailsByTransactionIdribsd2 {
             locktime,
             size,
