+++ src/models/get_transaction_details_by_transaction_idribsd_vin.rs
@@ -22,7 +22,7 @@ pub struct GetTransactionDetailsByTransactionIdribsdVin {
     pub script_sig: Box<crate::models::GetTransactionDetailsByTransactionIdribsdScriptSig>,
     /// Represents the script sequence number.
     #[serde(rename = "sequence")]
-    pub sequence: i32,
+    pub sequence: i64,
     /// Represents the reference transaction identifier.
     #[serde(rename = "txid", skip_serializing_if = "Option::is_none")]
     pub txid: Option<String>,
@@ -37,7 +37,7 @@ pub struct GetTransactionDetailsByTransactionIdribsdVin {
 }
 
 impl GetTransactionDetailsByTransactionIdribsdVin {
-    pub fn new(addresses: Vec<String>, script_sig: crate::models::GetTransactionDetailsByTransactionIdribsdScriptSig, sequence: i32, txinwitness: Vec<String>, value: String) -> GetTransactionDetailsByTransactionIdribsdVin {
+    pub fn new(addresses: Vec<String>, script_sig: crate::models::GetTransactionDetailsByTransactionIdribsdScriptSig, sequence: i64, txinwitness: Vec<String>, value: String) -> GetTransactionDetailsByTransactionIdribsdVin {
         GetTransactionDetailsByTransactionIdribsdVin {
             addresses,
             coinbase: None,
