+++ src/models/get_transaction_details_by_transaction_idribsz_vin.rs
@@ -19,7 +19,7 @@ pub struct GetTransactionDetailsByTransactionIdribszVin {
     pub script_sig: Box<crate::models::GetTransactionDetailsByTransactionIdribszScriptSig>,
     /// Represents the script sequence number.
     #[serde(rename = "sequence")]
-    pub sequence: i32,
+    pub sequence: i64,
     /// Represents the reference transaction identifier.
     #[serde(rename = "txid")]
     pub txid: String,
@@ -34,7 +34,7 @@ pub struct GetTransactionDetailsByTransactionIdribszVin {
 }
 
 impl GetTransactionDetailsByTransactionIdribszVin {
-    pub fn new(addresses: Vec<String>, script_sig: crate::models::GetTransactionDetailsByTransactionIdribszScriptSig, sequence: i32, txid: String, txinwitness: Vec<String>, value: String, vout: i32) -> GetTransactionDetailsByTransactionIdribszVin {
+    pub fn new(addresses: Vec<String>, script_sig: crate::models::GetTransactionDetailsByTransactionIdribszScriptSig, sequence: i64, txid: String, txinwitness: Vec<String>, value: String, vout: i32) -> GetTransactionDetailsByTransactionIdribszVin {
         GetTransactionDetailsByTransactionIdribszVin {
             addresses,
             script_sig: Box::new(script_sig),
