+++ src/models/get_wallet_transaction_details_by_transaction_idribsd2.rs
@@ -16,7 +16,7 @@
 pub struct GetWalletTransactionDetailsByTransactionIdribsd2 {
     /// Represents the time at which a particular transaction can be added to the blockchain.
     #[serde(rename = "locktime")]
-    pub locktime: i32,
+    pub locktime: i64,
     /// Represents the total size of this transaction.
     #[serde(rename = "size")]
     pub size: i32,
@@ -33,7 +33,7 @@ pub struct GetWalletTransactionDetailsByTransactionIdribsd2 {
 
 impl GetWalletTransactionDetailsByTransactionIdribsd2 {
     /// Dash
-    pub fn new(locktime: i32, size: i32, version: i32, vin: Vec<crate::models::GetWalletTransactionDetailsByTransactionIdribsd2Vin>, vout: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vout>) -> GetWalletTransactionDetailsByTransactionIdribsd2 {
+    pub fn new(locktime: i64, size: i32, version: i32, vin: Vec<crate::models::GetWalletTransactionDetailsByTransactionIdribsd2Vin>, vout: Vec<crate::models::GetTransactionDetailsByTransactionIdribsd2Vout>) -> GetWalletTransactionDetailsByTransactionIdribsd2 {
         GetWalletTransactionDetailsByTransactionIdribsd2 {
             locktime,
             size,
